requirements.txt
Django
djangorestframework
python-decouple
psycopg2-binary
#  removed in jenkis -    psycopg2

